requirements.txt
appdirs==1.4.3
Django==1.8.17
django-filter==1.0.1
djangorestframework==3.6.2
Markdown==2.6.8
packaging==16.8
pyparsing==2.2.0
requests==2.13.0
six==1.10.0
